pyproject.toml
[project]
name = "olfactorynet"
version = "0.1.0"
description = "GNN-based molecular odor prediction system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
    "pyrfume>=0.19.0",
    "rdkit>=2023.9.1",
    "hydra-core>=1.3.0",
    "torchmetrics>=1.2.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
logging = [
    "wandb>=0.15.0",
]
sweep = [
    "optuna>=3.3.0",
    "hydra-optuna-sweeper>=1.2.0",
]
ui = [
    "gradio>=4.0.0",
]
all = [
    "olfactorynet[dev,logging,sweep,ui]",
]

[project.scripts]
olfactorynet-train = "scripts.train:main"
olfactorynet-predict = "scripts.predict:main"
olfactorynet-app = "scripts.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/olfactorynet"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
